backend/app/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.1
sqlalchemy==2.0.20
databases==0.6.1
pymysql==1.0.3
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==1.3.1
python-dotenv==1.0.0
